% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_parametric_term.R
\name{evaluate_parametric_term}
\alias{evaluate_parametric_term}
\title{Evaluate a parametric term}
\usage{
evaluate_parametric_term(gam_model = NULL, pterm = NULL)
}
\arguments{
\item{gam_model}{A \code{gam} object model produced after fitting data from the \code{mgcv::gam()} function.}

\item{pterm}{A sting that names a parametric term to be evaluated}
}
\value{
A data frame of the evaluated parametric term
}
\description{
Evaluate a parametric term and return its values
}
\author{
Rick Dean
}
