% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_parametric_terms}
\alias{get_parametric_terms}
\title{get parametric terms of model}
\usage{
get_parametric_terms(gam_model)
}
\arguments{
\item{gam_model}{An object of class \dQuote{gam} produced by \code{mgcv::gam()}}
}
\description{
get parametric terms of model
}
