% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_vcov}
\alias{get_vcov}
\title{Get vcov ?}
\usage{
get_vcov(gam_model)
}
\arguments{
\item{gam_model}{An object of class \dQuote{gam} produced by \code{mgcv::gam()}}
}
\description{
Get vcov ?
}
