% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gam_1d.R
\name{plot_gam_1d}
\alias{plot_gam_1d}
\title{Plot smooth parameters estimated from Generalized Additive Models (GAM) based on ggplot2}
\usage{
plot_gam_1d(
  gam_model = NULL,
  select_smooth_terms = NULL,
  columns = 2,
  col_width = 4,
  row_height = 4,
  n_points = 100,
  se = TRUE,
  se_fill = NA,
  se_color = "red",
  title = NULL,
  subtitle = NULL,
  center_titles = FALSE,
  x_limits = NULL,
  x_major_breaks = waiver(),
  y_limits = NULL,
  y_major_breaks = waiver(),
  line_size = 1,
  line_color = "black"
)
}
\arguments{
\item{gam_model}{A \code{gam} one dimensional model produced after fitting data from the \code{mgcv::gam()} function.}

\item{select_smooth_terms}{A string vector of one or more data variables from the GAM model to be plotted.
If \code{NULL} the all the smooth parameters are plotted.}

\item{columns}{An integer that sets the number of columns for the multi-panel display.}

\item{col_width}{An integer that sets the width of each plot column in inches.}

\item{row_height}{An integer that sets the height of each plot column in inches.}

\item{n_points}{An integer that sets the number of points used for each 1-d plot.}

\item{se}{A logical which if TRUE upper and lower lines are plotted at 2 standard errors above and below the
the smooth being plotted.}

\item{se_fill}{A string that sets the fill color between the standard error lines.}

\item{se_color}{A string that sets the color of the standard error lines.}

\item{title}{A string that sets the plot title.}

\item{subtitle}{A string that sets the plot subtitle.}

\item{center_titles}{A logical which if \code{TRUE} centers both the \code{title} and \code{subtitle}.}

\item{x_limits}{Depending on the class of \code{aes_x}, a numeric/Date/POSIXct 2 element vector that sets the minimum
and maximum for the x axis.}

\item{x_major_breaks}{A numeric vector that defines the exact major tic locations along the x axis for all
smooth parameters. If not set then each smooth parameter has their own x axis scaling using ggplot2 defaults. In
addition the argument may be assigned to a named list (e.g. \code{list(x1 = seq(-3, 4, 1.0), x3 = seq(-2, 2, 0.5))}) of
numeric vectors.}

\item{y_limits}{A numeric 2 element vector that sets the minimum and  maximum for the y axis for all smooth parameters.}

\item{y_major_breaks}{A numeric vector that defines the exact major tic locations along the y axis for all
smooth parameters. If not set then each smooth parameter has their own y axis scaling using ggplot2 defaults. In
addition the argument may be assigned to a named list (e.g. \code{list(x1 = seq(-3, 4, 1.0), x3 = seq(-2, 2, 0.5))}) of
numeric vectors.}

\item{line_size}{A numeric value that sets the thickness of lines if \code{connect} is TRUE.}

\item{line_color}{A string that sets the color of the lines if \code{connect} is TRUE.}
}
\value{
A ggplot2 plot object
}
\description{
From a one dimensional GAM model produced by \code{mgcv::gam()}, produces a ggplot2 based plot of
selected smooth parameters in a multi-panel display.
}
\author{
Rick Dean
}
