% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{term_index}
\alias{term_index}
\title{Get the indices of a smooth label from a GAM model produced by \code{mgcv::gam()}}
\usage{
term_index(smooth_term, gam_model)
}
\arguments{
\item{gam_model}{An object of class \dQuote{gam} produced by \code{mgcv::gam()}}

\item{smooth_label}{A string that is the label of smooth object of class \code{mgcv.smooth}}
}
\description{
Get the indices of a smooth label from a GAM model produced by \code{mgcv::gam()}
}
\author{
Rick Dean
}
