% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{term_smooth}
\alias{term_smooth}
\title{Get the term of a smooth object}
\usage{
term_smooth(smooth_obj)
}
\arguments{
\item{smooth_obj}{An object of class \dQuote{mgcv.smooth} produced by \code{mgcv::gam()}}
}
\description{
Get the term of a smooth object
}
\author{
Rick Dean
}
