% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{terms_gam}
\alias{terms_gam}
\title{Get the terms of all smooth objects from a GAM model produced by \code{mgcv::gam()}}
\usage{
terms_gam(gam_model)
}
\arguments{
\item{gam_model}{An object of class \dQuote{gam} produced by \code{mgcv::gam()}}
}
\description{
Get the terms of all smooth objects from a GAM model produced by \code{mgcv::gam()}
}
\author{
Rick Dean
}
